web/requirements.txt
flask
flask-cors
pyyaml
